assert_all_are_dirs {assertive.files}	R Documentation

Is the path a directory? Checks to see if the input path is a directory.

Description

Is the path a directory? Checks to see if the input path is a directory.

Usage

assert_all_are_dirs(x, severity = getOption("assertive.severity", "stop"))

assert_any_are_dirs(x, severity = getOption("assertive.severity", "stop"))

is_dir(x, .xname = get_name_in_parent(x))

Arguments

x
File paths.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_dir returns TRUE if and only if the input path is a directory that exists, as determined by file.info.

Examples

assert_all_are_dirs(R.home())

[Package assertive.files version 0.0-2 Index]