assert_all_are_executable_files {assertive.files}	R Documentation

Is the file accessible?

Description

Checks to see if the input files can be executed/read/written to.

Usage

assert_all_are_executable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_any_are_executable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_all_are_readable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_any_are_readable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_all_are_writable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_any_are_writable_files(x, warn_about_windows = TRUE,
  severity = getOption("assertive.severity", "stop"))

is_executable_file(x, warn_about_windows = TRUE,
  .xname = get_name_in_parent(x))

is_ex_file(x)

is_readable_file(x, warn_about_windows = TRUE,
  .xname = get_name_in_parent(x))

is_writable_file(x, warn_about_windows = TRUE,
  .xname = get_name_in_parent(x))

Arguments

x
Input to check.

warn_about_windows
Logical. If TRUE, then calling the function under Windows will throw a warning about the problems with file.access.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_executable_file wraps file.access, showing the names of the inputs in the answer. assert_is_executable_file returns nothing but throws an error if is_executable_file returns FALSE.

See Also

file.access.

Examples

files <- dir()
is_readable_file(files)
is_writable_file(files, warn_about_windows = FALSE)
is_executable_file(files, warn_about_windows = FALSE)

[Package assertive.files version 0.0-2 Index]