assert_all_are_libraries {assertive.files}	R Documentation

Is the directory a known R library?

Description

Checks to see if the input directories are known R libraries.

Usage

assert_all_are_libraries(x, severity = getOption("assertive.severity",
  "stop"))

assert_any_are_libraries(x, severity = getOption("assertive.severity",
  "stop"))

is_library(x, .xname = get_name_in_parent(x))

Arguments

x
Directory paths

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_library returns TRUE if and only if the input paths are known R package libraries. That is, they must be paths returned by .libPaths.

Note

Input paths are converted to character, and then normalized using normalizePaths.

Examples

is_library(c(.libPaths(), R.home()))

[Package assertive.files version 0.0-2 Index]